[Identity] Add container instance live test
