+++ sdk/identity/azure-identity/tests/integration/azure-kubernetes-service/app.py
@@ -23,6 +23,20 @@ def run_sync():
     for container in containers:
         print(container["name"])
 
+    if os.environ.get("IDENTITY_USER_DEFINED_IDENTITY_CLIENT_ID") and os.environ.get(
+        "IDENTITY_STORAGE_NAME_USER_ASSIGNED"
+    ):
+        credential = ManagedIdentityCredential(client_id=os.environ.get("IDENTITY_USER_DEFINED_IDENTITY_CLIENT_ID"))
+
+        client = BlobServiceClient(
+            account_url=f"https://{os.environ['IDENTITY_STORAGE_NAME_USER_ASSIGNED']}.blob.core.windows.net",
+            credential=credential,
+        )
+
+        containers = client.list_containers()
+        for container in containers:
+            print(container["name"])
+
     print(f"Successfully acquired token with ManagedIdentityCredential")
 
 
@@ -39,6 +53,24 @@ async def run_async():
 
     await client.close()
     await credential.close()
+
+    if os.environ.get("IDENTITY_USER_DEFINED_IDENTITY_CLIENT_ID") and os.environ.get(
+        "IDENTITY_STORAGE_NAME_USER_ASSIGNED"
+    ):
+        credential = AsyncManagedIdentityCredential(
+            client_id=os.environ.get("IDENTITY_USER_DEFINED_IDENTITY_CLIENT_ID")
+        )
+
+        client = AsyncBlobServiceClient(
+            account_url=f"https://{os.environ['IDENTITY_STORAGE_NAME_USER_ASSIGNED']}.blob.core.windows.net",
+            credential=credential,
+        )
+
+        async for container in client.list_containers():
+            print(container["name"])
+
+        await client.close()
+        await credential.close()
 
     print("Successfully acquired token with async ManagedIdentityCredential")
