+++ sdk/identity/azure-identity/tests/integration/test_azure_container_instance.py
@@ -9,41 +9,39 @@ import pytest
 from utils import run_command
 
 
-class TestAzureKubernetesServiceIntegration:
+class TestAzureContainerInstanceIntegration:
     @pytest.mark.live_test_only
     @pytest.mark.skipif(
         not os.environ.get("IDENTITY_LIVE_RESOURCES_PROVISIONED"), reason="Integration resources not provisioned."
     )
-    def test_azure_kubernetes(self):
+    def test_azure_container_instance(self):
 
         client_id = os.environ.get("IDENTITY_CLIENT_ID")
         client_secret = os.environ.get("IDENTITY_CLIENT_SECRET")
         tenant_id = os.environ.get("IDENTITY_TENANT_ID")
         resource_group = os.environ.get("IDENTITY_RESOURCE_GROUP")
-        aks_cluster_name = os.environ.get("IDENTITY_AKS_CLUSTER_NAME")
         subscription_id = os.environ.get("IDENTITY_SUBSCRIPTION_ID")
-        pod_name = os.environ.get("IDENTITY_AKS_POD_NAME", "python-test-app")
+
+        container_instance_name = os.environ.get("IDENTITY_CONTAINER_INSTANCE_NAME", "python-container-app")
 
         az_path = run_command(["which", "az"])
-        kubectl_path = run_command(["which", "kubectl"])
         run_command(
             [az_path, "login", "--service-principal", "-u", client_id, "-p", client_secret, "--tenant", tenant_id]
         )
         run_command([az_path, "account", "set", "--subscription", subscription_id])
-        run_command(
+
+        # Using "script" as a workaround for "az container exec" requiring a tty.
+        # https://github.com/Azure/azure-cli/issues/17530
+        command = (
+            f"{az_path} container exec -g {resource_group} -n {container_instance_name} --exec-command 'python /app.py'"
+        )
+        output = run_command(
             [
-                az_path,
-                "aks",
-                "get-credentials",
-                "--resource-group",
-                resource_group,
-                "--name",
-                aks_cluster_name,
-                "--overwrite-existing",
+                "script",
+                "-q",
+                "-c",
+                command,
             ]
         )
-        pod_output = run_command([kubectl_path, "get", "pods", "-o", "jsonpath='{.items[0].metadata.name}'"])
-        assert pod_name in pod_output
 
-        output = run_command([kubectl_path, "exec", pod_name, "--", "python3", "/app.py"])
         assert "Passed!" in output
