+++ sdk/identity/azure-identity/tests/integration/test_azure_vms.py
@@ -3,21 +3,10 @@
 # Licensed under the MIT License.
 # ------------------------------------
 import os
-import sys
+
 import pytest
-import subprocess
 
-
-def run_command(command, exit_on_failure=True) -> str:
-    try:
-        result = subprocess.check_output(command, stderr=subprocess.STDOUT).decode("utf-8").strip()
-        return result
-    except subprocess.CalledProcessError as ex:
-        result = ex.output.decode("utf-8").strip()
-        if exit_on_failure:
-            print(result)
-            sys.exit(1)
-        return result
+from utils import run_command
 
 
 class TestAzureVirtualMachinesIntegration:
